+++ SystemInfo/Processes/WinApi/PTAPI/ProcessthreadsapiDLL.cs
@@ -6,7 +6,7 @@ namespace SystemInfo.Processes.WinApi.PTAPI
     {
 
         [DllImport("kernel32.dll", SetLastError = true)]
-        protected static extern IntPtr OpenProcess(long dwDesiredAccess, bool bInheritHandle, uint dwProcessId);
+        protected static extern IntPtr OpenProcess(uint dwDesiredAccess, bool bInheritHandle, uint dwProcessId);
 
         [DllImport("kernel32.dll", SetLastError = true)]
         protected static extern bool TerminateProcess(IntPtr hProcess, uint uExitCode);
